:recycle: Refactor: change Package Structure
  - 패키지 구조 변경
