+++ crispin/src/main/java/study/crispin/member/application/service/MemberServiceImpl.java
@@ -5,7 +5,6 @@ import study.crispin.member.application.request.MemberRegistrationRequest;
 import study.crispin.member.application.request.MemberUpdateRequest;
 import study.crispin.member.domain.Member;
 import study.crispin.member.infrastructure.repository.MemberRepository;
-import study.crispin.member.presentation.port.MemberService;
 import study.crispin.member.presentation.response.MemberRegistrationResponse;
 import study.crispin.member.presentation.response.MemberUpdateResponse;
 import study.crispin.team.infrastructure.repository.TeamRepository;
